done part 2

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: rcaillon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 20:56:33 by rcaillon          #+#    #+#             */
-/*   Updated: 2018/11/08 21:06:13 by rcaillon         ###   ########.fr       */
+/*   Updated: 2018/11/09 19:58:33 by rcaillon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	size_t	s1_len;
 	size_t	s2_len;
 	char	*new;
-	int		i;
+	size_t	i;
 
 	s1_len = ft_strlen(s1);
 	s2_len = ft_strlen(s2);
